+++ examples/user_plugins/js/main.js
@@ -1,0 +1,2 @@
+console.log(require.user_plugin_feature());
+console.log(require.user_plugin_with_options_feature());