+++ examples/user_plugins/plugins/invalid_user_plugin.js
@@ -1,0 +1,8 @@
+/**
+ * User plugin
+ */
+
+/**
+ * @name sandbox
+ */
+var invalid;